requirements.txt
allure-pytest==2.14.3
allure-python-commons==2.14.3
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.3
colorama==0.4.6
execnet==2.1.1
Faker==37.4.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==25.0
pluggy==1.6.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PySocks==1.7.1
pytest==8.4.1
pytest-rerunfailures==15.1
pytest-xdist==3.8.0
python-dotenv==1.1.1
requests==2.32.5
selenium==4.34.2
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.8.0
wsproto==1.2.0
